refactor: collect publish logic to custom hook
