+++ src/domains/creation/reducers/publish/CreationPublishQuestionSetReducer.ts
@@ -9,7 +9,7 @@ import type {
 //
 //
 
-type CreationPublishQuestionSetState = UpdateQuestionSetApiRequest & {
+export type CreationPublishQuestionSetState = UpdateQuestionSetApiRequest & {
 	categories: QuestionSetCategoryApiResponse[];
 };
